+++ WebContent/js/components/JobFile.js
@@ -16,6 +16,7 @@ import { ContextMenu, MenuItem, ContextMenuTrigger } from 'react-contextmenu';
 import Description from '@material-ui/icons/Description';
 import { hideMenu } from 'react-contextmenu/modules/actions';
 import { fetchJobFile, getFileLabel, changeSelectedContent, downloadFile } from '../actions/content';
+import { selectFile, unselectAllJobFiles, unselectAllJobs, highlightSelected } from '../actions/jobNodes';
 
 
 class JobFile extends React.Component {
@@ -59,6 +60,9 @@ class JobFile extends React.Component {
         } else {
             dispatch(fetchJobFile(job.get('jobName'), job.get('jobId'), file.label, file.id));
         }
+        dispatch(unselectAllJobFiles());
+        dispatch(unselectAllJobs());
+        dispatch(selectFile(job.get('jobId'), file.label));
     }
 
     refreshFile() {
@@ -82,6 +86,13 @@ class JobFile extends React.Component {
     hideContextMenu() {
         hideMenu();
         this.setState({ menuVisible: false });
+    }
+
+    handleContextMenu() {
+        const { dispatch, file } = this.props;
+        if (file.selectionType !== 'selected') {
+            dispatch(highlightSelected());
+        }
     }
 
     handleKeyDown(e) {
@@ -109,6 +120,7 @@ class JobFile extends React.Component {
 
     renderJobFileMenu() {
         const { job, file } = this.props;
+
         const menuItems = [
             <MenuItem onClick={this.downloadJobFile} key="download" >
                 <u>D</u>ownload
@@ -151,11 +163,16 @@ class JobFile extends React.Component {
                         <span
                             className="content-link"
                             onClick={() => { this.openFile(); }}
+                            onContextMenu={() => { this.handleContextMenu(); }}
                             onKeyDown={this.handleKeyDown}
                             tabIndex="0"
                             role="treeitem"
                             aria-level="2"
                             aria-haspopup={true}
+                            style={this.state.menuVisible ? { border: '1px solid #333333' }
+                                : file.selectionType === 'selected' ? { background: '#dedede', border: '1px solid #333333' }
+                                    : file.selectionType === 'highlighted' ? { background: '#dedede' }
+                                        : null}
                         >
                             <Description className="node-icon" />
                             <span className="job-file-label">{file.label}</span>
@@ -172,6 +189,7 @@ JobFile.propTypes = {
     showDD: PropTypes.string,
     content: PropTypes.instanceOf(List),
     file: PropTypes.shape({
+        selectionType: PropTypes.string.isRequired,
         label: PropTypes.string.isRequired,
         id: PropTypes.number.isRequired,
     }),