+++ WebContent/js/reducers/jobNodes.js
@@ -17,6 +17,9 @@ import {
     TOGGLE_JOB,
     INVERT_JOB_SELECT_STATUS,
     UNSELECT_ALL_JOBS,
+    UNSELECT_ALL_JOBS_FILES,
+    HIGHLIGHT_SELECTED,
+    SELECT_FILE,
     REQUEST_JOB_FILES,
     RECEIVE_JOB_FILES,
     RECEIVE_PURGE_JOB,
@@ -39,7 +42,7 @@ function extractJobs(jobs) {
             returnCode: job.retcode,
             status: job.status,
             isToggled: false,
-            isSelected: false,
+            selectionType: '',
             files: List(),
         };
     });
@@ -54,7 +57,7 @@ function extractJob(job) {
             returnCode: job.retcode,
             status: job.status,
             isToggled: false,
-            isSelected: false,
+            selectionType: '',
             files: List(),
         }),
     ]);
@@ -73,12 +76,55 @@ function toggleJob(jobs, jobId) {
 
 function invertJobSelectStatus(jobs, jobId) {
     const jobKey = findKeyOfJob(jobs, jobId);
-    return jobs.get(jobKey).set('isSelected', !jobs.get(jobKey).get('isSelected'));
+    return jobs.get(jobKey).set('selectionType', jobs.get(jobKey).get('selectionType') !== '' ? '' : 'selected');
 }
 
 function unselectAllJobs(jobs) {
     return jobs.map(job => {
-        return job.set('isSelected', false);
+        return job.set('selectionType', '');
+    });
+}
+
+function unselectAllJobFiles(jobs) {
+    return jobs.map(job => {
+        if (job.get('files')) {
+            job.set('files', job.get('files').forEach(jobFile => {
+                const file = jobFile;
+                file.selectionType = '';
+            }));
+        }
+        return job;
+    });
+}
+
+function highlightSelected(jobs) {
+    return jobs.map(job => {
+        if (job.get('selectionType') === 'selected') {
+            return job.set('selectionType', 'highlighted');
+        }
+        if (job.get('files')) {
+            job.set('files', job.get('files').forEach(jobFile => {
+                const file = jobFile;
+                if (file.selectionType === 'selected') {
+                    file.selectionType = 'highlighted';
+                }
+            }));
+        }
+        return job;
+    });
+}
+
+function selectFile(jobs, jobId, label) {
+    return jobs.map(job => {
+        if (job.get('files') && job.get('jobId') === jobId) {
+            job.set('files', job.get('files').forEach(jobFile => {
+                const file = jobFile;
+                if (file.label === label) {
+                    file.selectionType = 'selected';
+                }
+            }));
+        }
+        return job;
     });
 }
 
@@ -87,6 +133,7 @@ function extractJobFiles(jobFiles) {
         return {
             label: file.ddname,
             id: file.id,
+            selectionType: '',
         };
     });
 }
@@ -116,6 +163,18 @@ export default function JobNodes(state = INITIAL_STATE, action) {
         case UNSELECT_ALL_JOBS:
             return state.merge({
                 jobs: unselectAllJobs(state.get('jobs')),
+            });
+        case UNSELECT_ALL_JOBS_FILES:
+            return state.merge({
+                jobs: unselectAllJobFiles(state.get('jobs')),
+            });
+        case HIGHLIGHT_SELECTED:
+            return state.merge({
+                jobs: highlightSelected(state.get('jobs')),
+            });
+        case SELECT_FILE:
+            return state.merge({
+                jobs: selectFile(state.get('jobs'), action.jobId, action.label),
             });
         case REQUEST_JOB_FILES:
             return state.set('isFetching', true);